Put a condition to check if node already installed

+++ chef-repo/cookbooks/nodejs/recipes/default.rb
@@ -10,8 +10,10 @@
 # bash resource to install nodejs
 bash "nodejs" do
 	code <<-EOH
-		wget http://nodejs.org/dist/v0.12.5/node-v0.12.5-linux-x64.tar.gz
-		sudo tar -C /usr/local --strip-components 1 -xzf node-v0.12.5-linux-x64.tar.gz 
-		rm node-v0.12.5-linux-x64.tar.gz 
+		if ! [ -e /usr/local/bin/node ]; then
+			wget http://nodejs.org/dist/v0.12.5/node-v0.12.5-linux-x64.tar.gz
+			sudo tar -C /usr/local --strip-components 1 -xzf node-v0.12.5-linux-x64.tar.gz 
+			rm node-v0.12.5-linux-x64.tar.gz 
+		fi
 	EOH
 end
